refactor(api): extract shared keyset cursor helper for /v1
Move encodeCursor/decodeCursor into src/shared/utils/api/cursor.mjs (now that a
second endpoint needs them) and re-export them from the status projection so its
imports/tests are unchanged. Behavior-identical.
